Move config code to its own module
